requirements.txt
absl-py==0.9.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
future==0.18.2
google-auth==1.16.0
google-auth-oauthlib==0.4.1
grpcio==1.29.0
idna==2.9
joblib==0.15.1
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
numpy==1.18.4
oauthlib==3.1.0
pathlib==1.0.1
Pillow==7.1.2
progressbar2==3.51.3
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.4.0
requests==2.23.0
requests-oauthlib==1.3.0
rmate==1.0.3
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
threadpoolctl==2.1.0
torch==1.5.0
torchvision==0.6.0
urllib3==1.25.9
Werkzeug==1.0.1
